refactored simulation api with abstract classes
